% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dist_plotting.R
\name{binom_pmf}
\alias{binom_pmf}
\title{Plotting Binomial PMFs}
\usage{
binom_pmf(n, p, x_shaded = NULL, ...)
}
\description{
Plots and colours Binomial pmfs according to x-values given
}
\details{
Colouring not implemented, use \link{ggbinom_pmf} instead.
}
