% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ct}
\alias{ct}
\title{Cat Wrapper - One Newline}
\usage{
ct(..., file = "", sep = "", fill = FALSE, labels = NULL, append = FALSE)
}
\description{
A wrapper for cat that adds a newline at the
end of the output and uses an empty separator.
}
