% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ctt}
\alias{ctt}
\title{Cat Wrapper - Two Newlines}
\usage{
ctt(..., file = "", sep = "", fill = FALSE, labels = NULL, append = FALSE)
}
\description{
A wrapper for cat that adds a newline at the
end of the output and uses an empty separator.
}
